requirements.txt
# ============ CORE AGENT FRAMEWORK ============
pydantic-ai>=0.0.14            # Agent 1: Structured extraction (>=0.0.14 for griffe fix)
anthropic==0.42.0              # Agent 2: Claude SDK for conversation
openai==1.59.5                 # For GPT-4o-mini + embeddings

# ============ LANGGRAPH + LANGCHAIN (v4.0 Multi-Agent Orchestration) ============
langgraph==0.2.62              # Multi-agent state machine orchestration
langchain==0.3.15              # LangChain framework
langchain-anthropic==0.3.5     # Claude integration for LangChain
langchain-openai==0.3.1        # GPT integration for LangChain
langchain-community==0.3.10    # Community integrations

# ============ WEB FRAMEWORK ============
fastapi==0.115.6               # API backend
uvicorn[standard]==0.34.0      # ASGI server
python-multipart==0.0.20       # File upload support
httpx==0.28.1                  # Async HTTP client
slowapi==0.1.9                 # Rate limiting for FastAPI

# ============ DATABASE ============
supabase==2.22.0               # Supabase client (PostgreSQL + Auth + Storage)
pgvector==0.3.6                # PGVector extension for vector search
psycopg2-binary==2.9.10        # PostgreSQL adapter

# ============ DATA VALIDATION ============
pydantic[email]==2.10.4        # Type validation (with email-validator)
pydantic-settings==2.7.1       # Settings management

# ============ WHATSAPP INTEGRATION ============
requests==2.32.3               # HTTP requests for 360Dialog API
twilio==8.10.0                 # Twilio SDK for WhatsApp Business API

# ============ BACKGROUND JOBS ============
celery==5.4.0                  # Task queue
redis==5.2.1                   # Celery broker + rate limiting

# ============ WEB SCRAPING ============
playwright==1.49.1             # Headless browser for Seldenrijk inventory scraping

# ============ ENVIRONMENT ============
python-dotenv==1.0.1           # .env file support

# ============ TESTING ============
pytest==8.3.4                  # Testing framework
pytest-asyncio==0.25.2         # Async test support
pytest-cov==6.0.0              # Coverage reporting

# ============ DASHBOARD (REFLEX) ============
reflex==0.6.5                  # Modern Python web framework

# ============ MONITORING ============
sentry-sdk==2.19.2             # Error tracking
structlog==24.4.0              # Structured logging
prometheus-client==0.21.1      # Metrics
psutil==6.1.1                  # System metrics

# ============ DEVELOPMENT ============
black==24.10.0                 # Code formatting
ruff==0.8.5                    # Linting
mypy==1.14.1                   # Type checking
